requirements.txt
asn1crypto==0.24.0
asttokens==2.4.0
async-timeout==4.0.2
atomicwrites==1.1.5
attrs==18.1.0
bcrypt==3.1.4
blinker==1.4
cachetools==5.3.2
cairocffi==1.2.0
CairoSVG==2.7.1
certifi==2018.8.13
cffi==1.15.1
chardet==3.0.4
click==7.1.2
colorama==0.4.6
crontab==1.0.1
cryptography==38.0.3
cssselect2==0.7.0
dataclasses==0.6
defusedxml==0.7.1
et-xmlfile==1.0.1
executing==2.0.0
flasgger==0.9.0
Flask==2.0.3
Flask-Bcrypt==1.0.1
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.1
Flask-SQLAlchemy==2.5.1
freezegun==1.2.2
google-api-core==2.14.0
google-api-python-client==2.108.0
google-auth==2.23.4
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
googleapis-common-protos==1.61.0
greenlet==2.0.2
gunicorn==20.1.0
html5lib==1.1
httplib2==0.22.0
icecream==2.1.3
idna==2.7
importlib-metadata==4.8.3
inflect==6.0.4
itsdangerous==2.0.1
jdcal==1.4.1
Jinja2==3.0.3
jsonschema==3.2.0
MarkupSafe==2.0.1
misaka==2.1.0
mistune==2.0.5
more-itertools==4.3.0
neverbounce-sdk==4.3.0
oauthlib==3.2.2
openpyxl==3.0.10
packaging==21.3
Pillow==10.0.1
pluggy==0.7.1
protobuf==4.25.1
psycopg2-binary==2.9.9
py==1.5.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.18
pydantic==1.10.8
Pygments==2.16.1
PyJWT==2.4.0
pyOpenSSL==22.1.0
pyparsing==3.0.6
pyphen==0.11.0
pyrsistent==0.18.0
pytest==3.7.2
pytest-flask==0.10.0
pytest-voluptuous==1.0.2
python-dateutil==2.8.2
pytz==2020.1
PyYAML==6.0.1
redis==5.0.1
requests==2.23.0
requests-oauthlib==1.3.1
rq==1.15.1
rq-scheduler==0.13.1
rsa==4.9
sentry-sdk==1.6.0
six==1.16.0
SQLAlchemy==1.4.41
SQLAlchemy-Utils==0.39.0
tinycss2==1.1.1
twilio==7.9.2
typing_extensions==4.2.0
uritemplate==4.1.1
urllib3==1.23
voluptuous==0.13.1
WeasyPrint==52.5
webencodings==0.5.1
Werkzeug==2.0.3
wrapt==1.15.0
XlsxWriter==1.0.5
# zephony==0.60
zephony==1.3.1
zipp==3.6.0
